feat: add Islamic date conversion functions and integrate into getMoulidDate
